% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdaldem.R
\name{gdaldem}
\alias{gdaldem}
\title{R interface to GDAL's gdaldem utility}
\usage{
gdaldem(mode, input_dem, output_map, ..., of, compute_edges, alg, b, co, q, z,
  s, az, alt, combined, multidirectional, p, trigonometric, zero_for_flat,
  color_text_file = character(0), alpha, exact_color_entry,
  nearest_color_entry, dryrun = FALSE)
}
\value{
None. Called instead for its side effect.
}
\description{
This function provides an interface mirroring that of the GDAL
command-line app \code{gdaldem}. For a description of the
utility and the arguments that it takes, see the documentation at
\url{http://www.gdal.org/gdaldem.html}.
}
\examples{
\dontrun{
## Prepare file paths
td <- tempdir()
in_dem <- system.file("extdata/maunga.tif", package = "gdalUtilities")
out_slope  <- file.path(td, "slope.tif")
out_shade  <- file.path(td, "shade.tif")
out_aspect <- file.path(td, "aspect.tif")

## Apply DEM processing
gdaldem("slope", in_dem, out_slope)
gdaldem("shade", in_dem, out_shade)
gdaldem("aspect", in_dem, out_aspect)

## View results
if(require(rasterVis)) {
    lp <- function(f) {
        levelplot(raster(f), main = substitute(f),
                  margin = FALSE, colorkey = FALSE)
    }
    plot(lp(in_dem),     split = c(1,1,2,2))
    plot(lp(out_slope),  split = c(2,1,2,2), newpage = FALSE)
    plot(lp(out_shade),  split = c(1,2,2,2), newpage = FALSE)
    plot(lp(out_aspect), split = c(2,2,2,2), newpage = FALSE)
}
}
}
\author{
Josh O'Brien
}
