% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalmdimtranslate.R
\name{gdalmdimtranslate}
\alias{gdalmdimtranslate}
\title{Interface to GDAL's gdalmdimtranslate utility}
\usage{
gdalmdimtranslate(
  src_filename,
  dst_filename,
  ...,
  co,
  IF,
  of,
  array,
  group,
  subset,
  scaleaxes,
  oo,
  dryrun = FALSE
)
}
\arguments{
\item{src_filename}{Character. Path to a GDAL-supported readable
datasource.}

\item{dst_filename}{Character. Path to a GDAL-supported output
file.}

\item{...}{Here, a placeholder argument that forces users to
supply exact names of all subsequent formal arguments.}

\item{co, IF, of, array, group, subset, scaleaxes, oo}{See the GDAL
project's
\href{https://gdal.org/programs/gdalmdimtranslate.html}{gdalmdimtranslate
documentation} for details.}

\item{dryrun}{Logical (default \code{FALSE}). If \code{TRUE},
instead of executing the requested call to GDAL, the function
will print the command-line call that would produce the
equivalent output.}
}
\value{
Silently returns path to \code{dst_filename}.
}
\description{
This function provides an interface mirroring that of the GDAL
command-line app \code{gdalmdimtranslate}. For a description of the
utility and the arguments that it takes, see the documentation at
\url{https://gdal.org/programs/gdalmdimtranslate.html}.
}
\examples{
ff <- system.file("nc/cropped.nc", package = "sf")

}
\author{
Joshua O'Brien
}
