% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defunct.R
\docType{package}
\name{gdalUtilities-defunct}
\alias{gdalUtilities-defunct}
\alias{gRasterize}
\title{Defunct function(s) in the gdalUtilities package}
\usage{
gRasterize(...)
}
\arguments{
\item{...}{Function arguments}
}
\description{
These functions have been removed from this package.
}
\details{
\code{gRasterize} was removed due to its dependency on the
\pkg{raster} package, on which \pkg{gdalUtilities} no longer
Depends. The source for \code{gRasterize} may still be found (and
sourced, using \code{devtools::source_gist()}) at
\url{https://gist.github.com/JoshOBrien/7cf19b8b686e6d6230a78a1a9799883b}.
}
