\name{gdalUtilities-package}
\alias{gdalUtilities-package}
\alias{gdalUtilities}
\docType{package}
\title{\packageTitle{gdalUtilities}}
\description{\packageDescription{gdalUtilities}}
\details{
  The DESCRIPTION file: \packageDESCRIPTION{gdalUtilities}
  \packageIndices{gdalUtilities}

  Wrappers for 'GDAL' Utilities Executables.
}
\author{
  \packageAuthor{gdalUtilities}
  
  Maintainer: \packageMaintainer{gdalUtilities}
}
% Optionally other standard keywords, one per line,
% from the file KEYWORDS in the R documentation.
\keyword{package}

