% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gdalbuildvrt.R
\name{gdalbuildvrt}
\alias{gdalbuildvrt}
\title{R interface to GDAL's gdalbuildvrt utility}
\usage{
gdalbuildvrt(gdalfile, output.vrt, tileindex, resolution, te, tr, tap,
  separate, b, sd, allow_projection_difference, q, addalpha, hidenodata,
  srcnodata, vrtnodata, a_srs, r, input_file_list, overwrite,
  dryrun = FALSE)
}
\value{
None. Called instead for its side effect.
}
\description{
This function provides an interface mirroring that of the GDAL
command-line app \code{gdalbuildvrt}. For a description of the
utility and the arguments that it takes, see the documentation at
\url{https://gdal.org/programs/gdalbuildvrt.html}.
}
\examples{
## Prepare file paths
td <- tempdir()
out_vrt <- file.path(td, "out.vrt")
layer1 <-
    system.file("external/tahoe_lidar_bareearth.tif",
                package = "gdalUtils")
layer2 <-
    system.file("external/tahoe_lidar_highesthit.tif",
                package = "gdalUtils")

## Build VRT and check that it works
gdalbuildvrt(gdalfile = c(layer1, layer2), output.vrt = out_vrt)
gdalinfo(out_vrt)
}
\author{
Joshua O'Brien
}
