% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearblack.R
\name{nearblack}
\alias{nearblack}
\title{R interface to GDAL's nearblack utility}
\usage{
nearblack(infile, o = infile, ..., of, co, white, color, near, nb,
  setalpha, setmask, q, dryrun = FALSE)
}
\value{
Silently returns path to \code{datasetname}.
}
\description{
This function provides an interface mirroring that of the GDAL
command-line app \code{nearblack}. For a description of the
utility and the arguments that it takes, see the documentation at
\url{https://gdal.org/programs/nearblack.html}.
}
\examples{
\dontrun{
td <- tempdir()
a_rast <- file.path(td, "a.tif")
b_rast <- file.path(td, "b.tif")
file.copy(system.file("extdata/tahoe.tif", package = "gdalUtilities"),
          a_rast)
file.copy(system.file("extdata/tahoe.tif", package = "gdalUtilities"),
          b_rast)
nearblack(a_rast, b_rast, of = "GTiff")
nearblack(a_rast, of = "GTiff", near = 100)
}
}
\author{
Joshua O'Brien
}
