% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ogr2ogr.R
\name{ogr2ogr}
\alias{ogr2ogr}
\title{R interface to GDAL's ogr2ogr utility}
\usage{
ogr2ogr(src_datasource_name, dst_datasource_name, ..., layer, f, append,
  overwrite, update, select, progress, sql, dialect, where, skipfailures, spat,
  spat_srs, geomfield, dsco, lco, nln, nlt, dim, a_srs, t_srs, s_srs,
  preserve_fid, fid, oo, doo, gt, ds_transaction, clipsrc, clipsrcsql,
  clipsrclayer, clipsrcwhere, clipdst, clipdstsql, clipdstlayer, clipdstwhere,
  wrapdateline, datelineoffset, simplify, segmentize, fieldTypeToString,
  mapFieldType, unsetFieldWidth, splitlistfields, maxsubfields,
  explodecollections, zfield, gcp, order, tps, fieldmap, addfields,
  relaxedFieldNameMatch, forceNullable, unsetDefault, unsetFid, nomd, mo,
  dryrun = FALSE)
}
\value{
None. Called instead for its side effect.
}
\description{
This function provides an interface mirroring that of the GDAL
command-line app \code{ogr2ogr}.  For a description of the
utility and the arguments that it takes, see the documentation at
\url{http://www.gdal.org/ogr2ogr.html}.
}
\examples{
\dontrun{
## Prepare file paths
td <- tempdir()
in_shp <- system.file("external/lux.shp",
                      package = "raster")
out_merc <- file.path(td, "mercator.shp")
out_utm <- file.path(td, "utm.shp")

## Reproject to 'WGS 84/World Mercator'
## https://en.wikipedia.org/wiki/Mercator_projection
ogr2ogr(in_shp, out_merc, t_srs = "EPSG:3395")
## Reproject to 'Lambert conformal conic projection'
## https://en.wikipedia.org/wiki/Lambert_conformal_conic_projection
ogr2ogr(in_shp, out_utm, t_srs = "EPSG:42304")

if(require(raster)) {
    par(mfcol = c(1,2))
    plot(shapefile(out_merc), main = "WGS 84",
         border = "darkgrey", col = gray.colors(12))
    plot(shapefile(out_utm), main = "LCC",
         border = "darkgrey", col = gray.colors(12))
}
}
}
\author{
Josh O'Brien
}
